requirements.txt
# Enhanced LCD Game Engine Dependencies
# Core dependencies
Pillow>=9.0.0
RPLCD>=1.3.0
smbus2>=0.4.0
numpy>=1.21.0

# Input handling
gpiozero>=2.0.0

# Optional: For keyboard input and development
pygame>=2.0.0

# Optional: For enhanced features
# pyserial>=3.5  # For serial communication if needed
# adafruit-circuitpython-busdevice>=5.0.0  # For additional I2C devices 